+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Current
 
 - Make :class:`Namespace` behave like :class:`~flask.Blueprint` for :class:`~flask.Flask`.
 - Deprecated ``parser`` and ``body`` parameters for ``expect`` in :meth:`~Namespace.doc` decorator.
+- Added the :meth:`~Namespace.param` decorator
 
 0.8.6 (2015-12-26)
 ------------------